# docs/swagger.yaml
definitions:
...
  handlers.CreateReviewRequest:
    properties:
      comment:
        maxLength: 1000
        minLength: 10
        type: string
      rating:
        maximum: 5
        minimum: 1
        type: integer
      school_id:
        type: integer
    required:
    - comment
    - rating
    - school_id
    type: object
...
  handlers.ModerateReviewRequest:
    properties:
      notes:
        type: string
      status:
        allOf:
        - $ref: '#/definitions/models.ReviewStatus'
        enum:
        - approved
        - rejected
    required:
    - status
    type: object
...
  models.ReviewStatus:
    enum:
    - pending
    - approved
    - rejected
    type: string
    x-enum-varnames:
    - ReviewPending
    - ReviewApproved
    - ReviewRejected
...
info:
  contact: {}
paths:
...
  /admin/reviews/{review_id}/moderate:
    put:
      consumes:
      - application/json
      description: Allows an admin to approve or reject a pending review.
      parameters:
      - description: Review ID
        in: path
        name: review_id
        required: true
        type: integer
      - description: Moderation details (status and optional notes)
        in: body
        name: moderation
        required: true
        schema:
          $ref: '#/definitions/handlers.ModerateReviewRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Review moderated successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request or invalid review ID/request body
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden - only administrators can moderate reviews
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Review not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Moderate a review
      tags:
      - admin
      - reviews
  /admin/reviews/pending:
    get:
      description: Retrieves all reviews that are currently in 'pending' status, awaiting
        admin moderation.
      produces:
      - application/json
      responses:
        "200":
          description: List of pending reviews
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden - only administrators can view pending reviews
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get pending reviews
      tags:
      - admin
      - reviews
...
  /reviews:
    post:
      consumes:
      - application/json
      description: Creates a new review for a school (parents and educators only).
        The review will be pending approval by an admin.
      parameters:
      - description: Review information
        in: body
        name: review
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateReviewRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Review created successfully and pending approval
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request or validation error
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden - only parents and educators can create reviews
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: School not found
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: User has already reviewed this school
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a new review
      tags:
      - reviews
  /reviews/{review_id}:
    delete:
      description: Allows an authenticated user to delete their own review.
      parameters:
      - description: Review ID
        in: path
        name: review_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Review deleted successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad request or invalid review ID
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden - user can only delete their own reviews
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Review not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete a review
      tags:
      - reviews
    put:
      consumes:
      - application/json
      description: Allows an authenticated user to update their own review. The updated
        review will be reset to pending status.
      parameters:
      - description: Review ID
        in: path
        name: review_id
        required: true
        type: integer
      - description: Updated review information
        in: body
        name: review
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateReviewRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Review updated successfully and pending approval
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request or invalid review ID/request body
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden - user can only update their own reviews
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Review not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update a review
      tags:
      - reviews
  /reviews/me:
    get:
      description: Retrieves all reviews submitted by the authenticated user.
      produces:
      - application/json
      responses:
        "200":
          description: List of reviews by the current user
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get current user's reviews
      tags:
      - reviews
      - users
  /schools/{school_id}/reviews:
    get:
      description: Retrieves all approved reviews for a specific school, along with
        the average rating and total review count.
      parameters:
      - description: School ID
        in: path
        name: school_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of reviews with average rating
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request or invalid school ID
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: School not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get school reviews
      tags:
      - reviews
      - schools
...
swagger: "2.0"